Separate soccer logic into 4 files with 3 interactors.
